feat: update collection on-demand
Each user has x collections, assume that in each collection has more
than one word.

Instead of rendering all x collections at once, it will only render the
one that is currently being in use.

That is, we tell react to render the collection that the user wants to
see or add a new word to that collection.

This change, also made the use of useSWR more consistent as seen now
that each collection endpoint has its own set of words, so we can
further mutate it on DB operations.